Cleaned the project. Removed a few Log statements. Small changes to the comments.
Added the extra files for the sync adapter. Currently functions as with the service and clears the data on each sync. Alteration of the ContentProvider and how the articles are stored could be used to only update the current elements. An additional tag for changes or stale data could be added to the database.

Changed the animation for the sharedElementTransition. When returning to the MainActivity(ArticleListActivity), the animation usually should be postponed to allow Picasso to load the thumbnail image. More details in comments inside ArticleDetailActivity's SharedElementCallback() method.

Changed the NavUtils method for the toolbar inside the DetailActivity.
Now the activity is started for result. Using the toolbar UP button returns to the MainActivity (ArticleListActivity) returning the position to which the RecyclerView should scroll to.
Alternative method would be to use PreferenceManager.

TODO: A bug within the Fragments of the ViewPager inside the startPostponedEnterTransition() method- getActivity sometimes causes a crash while swiping between fragments. Multiple ways to fix the issue- find the most suitable one.